Move test helpers to alternate module to avoid running the same tests more than once
