+++ src/lib/dynamodb.test.ts
@@ -1,4 +1,5 @@
 import * as dynamodb from './dynamodb'
+import { _getServiceUsageQuotaItem, _updateServiceUsageQuota } from './dynamodb_test_helpers.test'
 
 describe('mergeEstateId', () => {
   test('it works', async () => {
@@ -103,69 +104,54 @@ describe('mergeEstateId', () => {
 
 describe('getEstateIdForAddress', () => {
   test('it returns different IDs for different buildings and addresses without buildings', async () => {
-    const id1 = await dynamodb.store({
-      rawAddress: "東京都千代田区永田町１丁目７−１",
-      address: "東京都千代田区永田町一丁目7-1",
-      tileXY: "3726576/1649777",
-      zoom: 22,
-      prefCode: "11"
-    })
+    const address = `getEstateIdForAddress - return different ID`
+    const [
+      id1,
+      id2,
+      id3,
+    ] = await Promise.all([
+      dynamodb.store({
+        rawAddress: "東京都千代田区永田町１丁目７−１",
+        address,
+        tileXY: "3726576/1649777",
+        zoom: 22,
+        prefCode: "11"
+      }),
+      dynamodb.store({
+        rawAddress: "東京都千代田区永田町１丁目７−１",
+        address,
+        rawBuilding: "国会議事堂A棟",
+        building: "国会議事堂A棟",
+        tileXY: "3726576/1649777",
+        zoom: 22,
+        prefCode: "11"
+      }),
+      dynamodb.store({
+        rawAddress: "東京都千代田区永田町１丁目７−１",
+        address,
+        rawBuilding: "国会議事堂B棟",
+        building: "国会議事堂B棟",
+        tileXY: "3726576/1649777",
+        zoom: 22,
+        prefCode: "11"
+      }),
+    ])
 
-    const id2 = await dynamodb.store({
-      rawAddress: "東京都千代田区永田町１丁目７−１",
-      address: "東京都千代田区永田町一丁目7-1",
-      rawBuilding: "国会議事堂A棟",
-      building: "国会議事堂A棟",
-      tileXY: "3726576/1649777",
-      zoom: 22,
-      prefCode: "11"
-    })
+    const [
+      queryResp1,
+      queryResp2,
+      queryResp3,
+    ] = await Promise.all([
+      dynamodb.getEstateIdForAddress(address),
+      dynamodb.getEstateIdForAddress(address, "国会議事堂A棟"),
+      dynamodb.getEstateIdForAddress(address, "国会議事堂B棟"),
+    ])
 
-    const id3 = await dynamodb.store({
-      rawAddress: "東京都千代田区永田町１丁目７−１",
-      address: "東京都千代田区永田町一丁目7-1",
-      rawBuilding: "国会議事堂B棟",
-      building: "国会議事堂B棟",
-      tileXY: "3726576/1649777",
-      zoom: 22,
-      prefCode: "11"
-    })
-
-    const queryResp1 = await dynamodb.getEstateIdForAddress("東京都千代田区永田町一丁目7-1")
     expect(queryResp1).toMatchObject(id1)
-
-    const queryResp2 = await dynamodb.getEstateIdForAddress("東京都千代田区永田町一丁目7-1", "国会議事堂A棟")
     expect(queryResp2).toMatchObject(id2)
-
-    const queryResp3 = await dynamodb.getEstateIdForAddress("東京都千代田区永田町一丁目7-1", "国会議事堂B棟")
     expect(queryResp3).toMatchObject(id3)
   })
 })
-
-export const _updateServiceUsageQuota = async ( usageKey:string, updateRequestCount:number ) => {
-
-  const updateItemInput: AWS.DynamoDB.DocumentClient.UpdateItemInput = {
-    TableName: process.env.AWS_DYNAMODB_API_KEY_TABLE_NAME,
-    Key: { apiKey : usageKey },
-    UpdateExpression: 'SET #c = :c',
-    ExpressionAttributeNames: {
-      '#c': 'c',
-    },
-    ExpressionAttributeValues: {
-      ':c': updateRequestCount
-    }
-  }
-  await dynamodb.DB.update(updateItemInput).promise()
-}
-
-export const _getServiceUsageQuotaItem = async ( usageKey:string ) =>{
-  const getItemInput: AWS.DynamoDB.DocumentClient.GetItemInput = {
-    TableName: process.env.AWS_DYNAMODB_API_KEY_TABLE_NAME,
-    Key: { apiKey : usageKey },
-  }
-  const { Item: item } = await dynamodb.DB.get(getItemInput).promise()
-  return item
-}
 
 describe('checkServiceUsageQuota', () => {
